Restore use of initializer to define CHAT_*
Environment variables and Rails views are at two opposite ends
of the system: one very close to the environment, and the other
very close to the "outputs". It probably does not make a large
difference whether we wrap them or not in practice, but it
seems better to avoid referencing ENV vars directly in
views if we don't have to.